Extract identity baggage into a testable module + add tests
Move the per-process identity helpers and the propagator out of main.py
into pulse_otel/identity.py so they can be unit-tested without importing
the heavy main module, and add tests covering the seeded keys, dropping of
empty/placeholder values, and the propagator's baggage header. Also clarify
that only non-identity inbound baggage passes through. Addresses Copilot
review on #76.
