+++ pulse_otel/__init__.py
@@ -1,3 +1,4 @@
-from pulse_otel.main import Pulse, CustomFileSpanExporter, FileLogExporter, pulse_agent, pulse_tool, observe, traced_function, setup_json_file_logger, seed_identity_baggage
+from pulse_otel.main import Pulse, CustomFileSpanExporter, FileLogExporter, pulse_agent, pulse_tool, observe, traced_function, setup_json_file_logger
+from pulse_otel.identity import seed_identity_baggage
 from pulse_otel.util import is_s2_owned_app
 from pulse_otel.consts import BAGGAGE_ORG, BAGGAGE_PROJECT, BAGGAGE_NOVA_ID, BAGGAGE_NOVA_TYPE, BAGGAGE_NOVA_NAME
